Ejercicio 2 hasta 10)
